+++ src/main/kotlin/nl/bryanderidder/regexrenamefiles/icons/RegexRenameFilesIcons.kt
@@ -1,0 +1,9 @@
+package nl.bryanderidder.regexrenamefiles.icons
+
+import com.intellij.openapi.util.IconLoader
+
+object RegexRenameFilesIcons {
+
+    @JvmField
+    val ActionIcon = IconLoader.getIcon("/icons/actionIcon.svg", RegexRenameFilesIcons::class.java)
+}